backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
gunicorn>=21.2,<22.0
black>=23.12,<25.0
flake8>=7.0,<8.0
pytest>=7.4,<9.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0
Faker>=22.0,<30.0
